compiled_clone to utils
